Move TorrcConfigGenerator to tor.common module
