+++ network/tor/tor-local-network/src/main/java/bisq/tor/local_network/torrc/DirectoryAuthorityTorrcGenerator.java
@@ -17,6 +17,7 @@
 
 package bisq.tor.local_network.torrc;
 
+import bisq.network.tor.common.TorrcConfigGenerator;
 import bisq.tor.local_network.TorNode;
 
 import java.util.Map;
